requirements.txt
gunicorn
flask
flask-login
flask-sqlalchemy
flask-bcrypt
pymysql
python-dotenv